Updated Status Messages
Made readable.

Reason for changing the order of the app.use and the console.log functions for routing and static directories is because if it were to fail, users shouldn't have to look back at the configuration file to figure out what router or directory it failed at.

Now they'd correctly see that the one that failed is the last one stated in the terminal.